# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediConnect - Patient and Doctor Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /users/me/complete-onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete user onboarding (set role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteOnboardingBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Bad request
...
  /users/my-patients:
    get:
      operationId: getMyPatients
      tags: [users]
      summary: Get list of patients for the current doctor
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientProfile"
        "403":
          description: Forbidden - not a doctor
  /users/add-patient:
    post:
      operationId: addPatient
      tags: [users]
      summary: Doctor adds a patient to their list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPatientBody"
      responses:
        "200":
          description: Success
        "403":
          description: Forbidden - not a doctor
        "404":
          description: Patient not found
  /users/my-doctor:
    get:
      operationId: getMyDoctor
      tags: [users]
      summary: Get the doctor assigned to the current patient
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorProfile"
        "404":
          description: No doctor assigned

  # Messaging Endpoints
...
  /messages/conversations/{contactId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Forbidden

  # Appointment Endpoints
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        role:
          type: ["string", "null"]
          enum: ["patient", "doctor", null]
        name:
          type: string
        email:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        conditions:
          type: ["string", "null"]
        qualifications:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - name
        - email
        - createdAt
    UpdateUserProfileBody:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        conditions:
          type: ["string", "null"]
        qualifications:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
    CompleteOnboardingBody:
      type: object
      properties:
        role:
          type: string
          enum: ["patient", "doctor"]
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        conditions:
          type: ["string", "null"]
        qualifications:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
      required:
        - role
        - name
    PatientProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        conditions:
          type: ["string", "null"]
      required:
        - id
        - clerkId
        - name
        - email
    DoctorProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        qualifications:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        nextAppointment:
          type: ["string", "null"]
      required:
        - id
        - clerkId
        - name
        - email
    AddPatientBody:
      type: object
      properties:
        patientClerkId:
          type: string
      required:
        - patientClerkId
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
        sentAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - senderId
        - receiverId
        - content
        - sentAt
        - isRead
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...